Deduplicate server creation logic between prod and dev server
